AI-Smart-Road-Monitoring-main/AI-Smart-Road-Monitoring-main/python_ai/requirements.txt
Flask
flask-cors
numpy
opencv-python
joblib
Pillow
pandas
scikit-learn==1.7.2
requests